+++ tests/unit/test_dependency_analyzer.py
@@ -5,6 +5,7 @@
 
 from unittest import TestCase
 
+import pytest
 from rpmfluff import SimpleRpmBuild
 from rpmfluff.yumrepobuild import YumRepoBuild
 
@@ -13,7 +14,7 @@ from rpmdeplint.repodata import Repo
 
 
 class TestDependencyAnalyzer(TestCase):
-    def test_repos(self):
+    def test_repos(self) -> None:
         lemon = SimpleRpmBuild("lemon", "1", "3", ["noarch"])
         lemon.add_provides("lemon-juice")
         lemon.add_provides("lemon-zest")
@@ -101,3 +102,45 @@ class TestDependencyAnalyzer(TestCase):
             )
             == 3
         )
+
+
+@pytest.mark.parametrize(
+    ("paths", "remove"),
+    [
+        pytest.param(
+            {
+                "/usr/lib/.build-id",
+                "/usr/lib/.build-id/df",
+                "/usr/lib/.build-id/0a",
+                "/usr/lib/debug/.build-id/ab",
+                "/usr/lib/debug",
+                "/usr/lib/debug/bin",
+                "/usr/lib/debug/sbin",
+                "/usr/lib/debug/lib",
+                "/usr/lib/debug/lib64",
+                "/usr/lib/debug/.dwz",
+                "/usr/lib/debug/usr",
+                "/usr/lib/debug/usr/bin",
+                "/usr/lib/debug/usr/sbin",
+                "/usr/lib/debug/usr/lib",
+                "/usr/lib/debug/usr/lib64",
+            },
+            True,
+            id="remove",
+        ),
+        pytest.param(
+            {
+                "/usr/lib/.build-id/0A",
+                "/usr/lib/.build-id/0g",
+                "/usr/lib/.build-id/0a/583dc4e283975c4e72c30104aba0541133e4e6",
+                "/usr/lib/not-build-id/34",
+                "/usr/lib",
+            },
+            False,
+            id="leave",
+        ),
+    ],
+)
+def test__remove_dirs_known_to_be_owned_by_many(paths: set[str], remove: bool) -> None:
+    expected = set() if remove else paths
+    assert DependencyAnalyzer._remove_dirs_known_to_be_owned_by_many(paths) == expected